server: Rename classes
- HackSoundServer -> src/server.py:Server
- HackSoundPlayer -> src/sound.py:Sound
  Log lines matching "player" have been updated to match the
  string "sound".

https://phabricator.endlessm.com/T25991
